backend/requirements.txt
Flask==2.2.5
openai==0.27.10
torch==2.0.1
transformers==4.34.0
Pillow==9.5.0
opencv-python==4.5.5.64
scikit-learn==1.3.0
nltk==3.8.1
numpy==1.25.2
